Include scala_library:java_sources in targets
Include java_sources in targets - the final target set returned by context.targets() was not considering targets that weren't specified in the "dependencies" of the scala_library. In this case, a target was specified using "scala_library:java_sources" which was being dropped. This in turn was causing idea not to load the project with the java sources it should have had.

I have added the logic in BuildGraph to get it working but it doesn't seem particularly elegant.

When adding java_sources we could attach a specific type to the target like "scala_java_sources_type" and then check for that during the target inclusion test instead of the extensive checking that I do now. Or folks who know BuildGraph might have better ideas - open to suggestions.

Testing Done:
CI passed locally. Idea project generated successfully with java_sources and compiles:

$./pants goal idea src/scala/com/pants/testproject/javasources::

And travis underway:
https://travis-ci.org/pantsbuild/pants/builds/30786335

Bugs closed: 390

Reviewed at https://rbcommons.com/s/twitter/r/733/